fix(control-plane): keep TASK-677 token scrub off desktop_backend.rs budget
v0.36.26 desktop-split gate forbids growing desktop_backend.rs above 6152 lines. Move the helper and tests to lib.rs; git spawn still scrubs via a same-line call.
